add stars rating to users

+++ db/migrate/20170405173810_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration[5.0]
   def change
     create_table :reviews do |t|
       t.belongs_to :user, index: true
-      t.float :rating
+      t.integer :rating
       t.text :comment
 
       t.timestamps
